# H3b-1050-web — 1-10-50 存量拆分(仅 apps/web)
Written by opencode (opencode-go/deepseek-v4-flash, variant max); Opus 5 终审通过。
