+++ apps/web/tests/msw/chat-handlers.ts
@@ -1,75 +1,34 @@
-import { readFileSync } from "node:fs";
-import { join } from "node:path";
 import { http, HttpResponse } from "msw";
 import type { HttpHandler } from "msw";
 import { TURNSTILE_HEADER } from "../../src/lib/turnstile/tokenStore";
+import { CHAT_URL, HEALTHZ_URL, chatStreamFixture, streamText } from "./chat-stream-base";
+import type { ChatStreamFixture, ChatStreamOptions } from "./chat-stream-base";
+import { sseResponse, SSE_HEADERS } from "./chat-sse";
 import { TEST_ORIGIN } from "./fixtures";
+
+export { CHAT_URL, HEALTHZ_URL } from "./chat-stream-base";
+export type { ChatStreamFixture, ChatStreamOptions } from "./chat-stream-base";
+export { chatStreamFixture, patchSessionId, streamText } from "./chat-stream-base";
+export type {
+  ControlledChatStream,
+  ControlledRecomputeStream,
+  FinalFramePatch,
+} from "./chat-recompute-handlers";
+export {
+  chatRecomputeControlledHandler,
+  chatRecomputeHandler,
+  chatStreamControlledHandler,
+  chatStreamHeldOpenHandler,
+  chatStreamPatchedHandler,
+  recomputeStreamFixture,
+  searchResultsPatch,
+} from "./chat-recompute-handlers";
 
 /**
  * Chat swimlane helpers. Stream bodies are REAL recordings replayed from
  * `apps/agent/tests/fixtures/chat_stream/*.sse` (E1's captures) — handlers
  * never hand-write AI SDK frames.
  */
-export const CHAT_URL = `${TEST_ORIGIN}/v1/chat`;
-export const HEALTHZ_URL = `${TEST_ORIGIN}/healthz`;
-
-/** Vitest runs with cwd = apps/web; the recordings live in the agent package. */
-const FIXTURE_DIR = join(process.cwd(), "..", "agent", "tests", "fixtures", "chat_stream");
-
-export type ChatStreamFixture = "search" | "clarify" | "error";
-
-export function chatStreamFixture(name: ChatStreamFixture): string {
-  return readFileSync(join(FIXTURE_DIR, `${name}.sse`), "utf8");
-}
-
-function sseBody(text: string, close: boolean): ReadableStream<Uint8Array> {
-  return new ReadableStream<Uint8Array>({
-    start(controller) {
-      controller.enqueue(new TextEncoder().encode(text));
-      if (close) controller.close();
-    },
-  });
-}
-
-export interface ChatStreamOptions {
-  /** Patch the recorded final frame's `session_id` (recordings capture null). */
-  readonly sessionId?: string;
-  /** Observe each request hitting the chat endpoint (headers assertions). */
-  readonly spy?: (request: Request) => void;
-  /** Corrupt the recorded final frame (type-invalid `success`) to probe schema guards. */
-  readonly malformedFinal?: boolean;
-  /** Serve this handler for a single request only (msw one-time handler). */
-  readonly once?: boolean;
-}
-
-function patchSessionId(text: string, sessionId?: string): string {
-  if (!sessionId) return text;
-  return text
-    .split("\n")
-    .map((line) => patchSessionIdLine(line, sessionId))
-    .join("\n");
-}
-
-/** The recordings omit the null `session_id`; inject it into full final frames. */
-function patchSessionIdLine(line: string, sessionId: string): string {
-  if (!line.startsWith('data: {"type":"data-response"')) return line;
-  const frame = JSON.parse(line.slice("data: ".length)) as { data: Record<string, unknown> };
-  if (!("success" in frame.data)) return line;
-  frame.data.session_id = sessionId;
-  return `data: ${JSON.stringify(frame)}`;
-}
-
-function corruptFinalFrame(text: string, malformed?: boolean): string {
-  if (!malformed) return text;
-  return text.replace('"success":true', '"success":"yep"');
-}
-
-function streamText(name: ChatStreamFixture, options: ChatStreamOptions): string {
-  return corruptFinalFrame(
-    patchSessionId(chatStreamFixture(name), options.sessionId),
-    options.malformedFinal,
-  );
-}
 
 export function chatStreamHandler(
   name: ChatStreamFixture,
@@ -183,188 +142,6 @@ export function chatStreamDropHandler(name: ChatStreamFixture): HttpHandler {
 /** Drops the connection before any frame arrives (D4 before-first-chunk). */
 export function chatStreamImmediateDropHandler(): HttpHandler {
   return http.post(CHAT_URL, () => new HttpResponse(droppingBody(""), { headers: SSE_HEADERS }));
-}
-
-export type FinalFramePatch = (envelope: Record<string, unknown>) => Record<string, unknown>;
-
-function patchFinalFrameLine(line: string, patch: FinalFramePatch): string {
-  if (!line.startsWith('data: {"type":"data-response"')) return line;
-  const frame = JSON.parse(line.slice("data: ".length)) as { data: Record<string, unknown> };
-  if (!("success" in frame.data)) return line;
-  frame.data = patch(frame.data);
-  return `data: ${JSON.stringify(frame)}`;
-}
-
-/**
- * Replay a recording with its final full envelope transformed. D-state
- * variants (D1/D2/D6/D9) are derived from the real capture this way until the
- * backend error-boundary hook (issue #272's other half) ships recordings of
- * the actual failure envelopes.
- */
-export function chatStreamPatchedHandler(
-  name: ChatStreamFixture,
-  patch: FinalFramePatch,
-  options: ChatStreamOptions = {},
-): HttpHandler {
-  const patched = streamText(name, options)
-    .split("\n")
-    .map((line) => patchFinalFrameLine(line, patch))
-    .join("\n");
-  return http.post(
-    CHAT_URL,
-    ({ request }) => {
-      options.spy?.(request);
-      return sseResponse(patched);
-    },
-    { once: options.once === true },
-  );
-}
-
-/** The E2 search-results envelope: derived from the real capture, like the
- * D-state variants, until a search_bangumi results recording ships. */
-export function searchResultsPatch(envelope: Record<string, unknown>): Record<string, unknown> {
-  const rows = [
-    { id: "p1", name: "宇治橋", latitude: 34.891, longitude: 135.807 },
-    { id: "p2", name: "京阪宇治駅", latitude: 34.911, longitude: 135.806 },
-    { id: "p3", name: "宇治神社", latitude: 34.9, longitude: 135.81 },
-  ];
-  return { ...envelope, intent: "search_bangumi", data: { results: { rows } } };
-}
-
-/** The real bypass step frames: `execute_selected_route` emits one
- * `plan_selected` running/done step pair (`test_selected_route.py`), which
- * `chat_stream._ToolPartTranslator` translates into exactly these tool
- * chunks. The UI must prove it SUPPRESSES them — a fixture without them
- * would certify the wrong tree (#461 review P1-1). */
-const PLAN_SELECTED_STEP_FRAMES = [
-  'data: {"type":"tool-input-start","toolCallId":"plan_selected-fixture","toolName":"plan_selected"}',
-  'data: {"type":"tool-input-available","toolCallId":"plan_selected-fixture","toolName":"plan_selected","input":{}}',
-  'data: {"type":"tool-output-available","toolCallId":"plan_selected-fixture","output":{"point_count":2}}',
-].join("\n\n");
-
-const START_STEP_FRAME = 'data: {"type":"start-step"}';
-
-/**
- * The `selected_point_ids` bypass stream (issue #273 S1.7 E2): the recorded
- * search stream with the agent-path tool frames replaced by the bypass's own
- * `plan_selected` step frames, and the route re-intended as `plan_selected`.
- */
-function toRecomputeStream(recording: string): string {
-  return recording
-    .split("\n")
-    .filter((line) => !line.startsWith('data: {"type":"tool-'))
-    .join("\n")
-    .replace(START_STEP_FRAME, `${START_STEP_FRAME}\n\n${PLAN_SELECTED_STEP_FRAMES}`)
-    .replaceAll('"intent":"plan_route"', '"intent":"plan_selected"');
-}
-
-/** Fixture self-guard: the recompute stream a test replays. Tests assert it
- * still carries the injected `plan_selected` step frames — deleting them
- * would silently re-certify the P1-1 false-green tree. */
-export function recomputeStreamFixture(): string {
-  return toRecomputeStream(chatStreamFixture("search"));
-}
-
-export function chatRecomputeHandler(options: ChatStreamOptions = {}): HttpHandler {
-  const recorded = toRecomputeStream(streamText("search", options));
-  return http.post(CHAT_URL, ({ request }) => {
-    options.spy?.(request);
-    return sseResponse(recorded);
-  });
-}
-
-export interface ControlledRecomputeStream {
-  readonly handler: HttpHandler;
-  /** Flush the final full envelope (and close); the skeleton shows until then. */
-  readonly releaseFinal: () => void;
-}
-
-/** The recompute stream held open before its final full envelope, so a test
- * can assert the skeleton state actually appeared (review P2-⑥). */
-export function chatRecomputeControlledHandler(): ControlledRecomputeStream {
-  const recorded = toRecomputeStream(chatStreamFixture("search"));
-  const splitAt = recorded.lastIndexOf('data: {"type":"data-response"');
-  let release: () => void = () => undefined;
-  const handler = http.post(CHAT_URL, () => {
-    const body = new ReadableStream<Uint8Array>({
-      start(controller) {
-        controller.enqueue(new TextEncoder().encode(recorded.slice(0, splitAt)));
-        release = () => {
-          flushTail(controller, recorded.slice(splitAt));
-        };
-      },
-    });
-    return new HttpResponse(body, { headers: SSE_HEADERS });
-  });
-  return {
-    handler,
-    releaseFinal: () => {
-      release();
-    },
-  };
-}
-
-/** Replays the recording up to (excluding) the first data-response frame and holds the stream open. */
-export function chatStreamHeldOpenHandler(name: ChatStreamFixture): HttpHandler {
-  const recorded = chatStreamFixture(name);
-  const head = recorded.slice(0, recorded.indexOf('data: {"type":"data-response"'));
-  return http.post(CHAT_URL, () => {
-    return sseResponse(head, { close: false });
-  });
-}
-
-export interface ControlledChatStream {
-  readonly handler: HttpHandler;
-  /** Flush the recorded final data-response frame (and close), if still open. */
-  readonly releaseFinal: () => void;
-}
-
-/** Streams the recording head, then lets the test release the final frame late. */
-export function chatStreamControlledHandler(
-  name: ChatStreamFixture,
-  sessionId: string,
-): ControlledChatStream {
-  const recorded = patchSessionId(chatStreamFixture(name), sessionId);
-  const splitAt = recorded.indexOf('data: {"type":"data-response"');
-  let release: () => void = () => undefined;
-  const handler = http.post(CHAT_URL, () => {
-    const body = new ReadableStream<Uint8Array>({
-      start(controller) {
-        controller.enqueue(new TextEncoder().encode(recorded.slice(0, splitAt)));
-        release = () => {
-          flushTail(controller, recorded.slice(splitAt));
-        };
-      },
-    });
-    return new HttpResponse(body, { headers: SSE_HEADERS });
-  });
-  return {
-    handler,
-    releaseFinal: () => {
-      release();
-    },
-  };
-}
-
-function flushTail(controller: ReadableStreamDefaultController<Uint8Array>, tail: string): void {
-  try {
-    controller.enqueue(new TextEncoder().encode(tail));
-    controller.close();
-  } catch {
-    // The consumer aborted the stream first: the late frame has nowhere to go.
-  }
-}
-
-const SSE_HEADERS = {
-  "content-type": "text/event-stream",
-  "x-vercel-ai-ui-message-stream": "v1",
-};
-
-function sseResponse(
-  text: string,
-  { close = true }: Readonly<{ close?: boolean }> = {},
-): HttpResponse<ReadableStream<Uint8Array>> {
-  return new HttpResponse(sseBody(text, close), { headers: SSE_HEADERS });
 }
 
 export const healthzOkHandler = http.get(HEALTHZ_URL, () =>
